Creacion del Servicio faltante, y avances de los controladores

+++ HorizonFutureVest.Web/ViewModels/RankingResultViewModel.cs
@@ -5,7 +5,7 @@ public class RankingResultViewModel
     public string NombrePais { get; set; } = null!;
     public string CodigoPais { get; set; } = null!;
     public decimal Puntaje { get; set; }
-    public int TasaRetornoEstimada  { get; set; }
+    public decimal TasaRetornoEstimada  { get; set; }
     public int Posicion { get; set; }
 
     public string PuntajeFormateado => $"{Puntaje:F2}";
